requirements.txt
fastapi
uvicorn
python-decouple
redis
motor
dnspython
starlette-discord